pyproject.toml
[project]
name = "wordlywise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic[email]>=2.11.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
    "alembic>=1.16.5",
    "python-jose>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "greenlet>=3.2.4",
    "openai>=1.108.0",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
]
